+++ docs/mdx-components.tsx
@@ -1,0 +1,7 @@
+import CodeBlock from "./components/CodeBlock";
+
+export function useMDXComponents() {
+  return {
+    pre: CodeBlock,
+  };
+}